DRY up multiprotocol writer
